Add script to launch odoo with additional checks
This is useful in a development environment where we're constantly launching and restarting Odoo.
On rare occasions, the Odoo server does not shut down gracefully and the port stays in use, Odoo does give out a warning regarding this but it gets buried underneath the other info messages and for some reason the server keeps running without a valid port.
This shell script checks if an Odoo/Python instance is utilizing port 8069, and promptly kills it.
Additionally, the Odoo server logs are passed through another output buffer whilst preserving format and colors.

One known issue with the script currently is that Signals are not passed correctly to the Odoo server and the graceful shutdown mechanism isn't working as of yet. This may cause data corruption...